Makefile improvements, adding toke role

+++ Makefile
@@ -1,27 +1,29 @@
 .PHONY: logo deploy docs_build restore develop docs_deploy lint
 
 # Deploy HomelabOS
-deploy: logo get_roles
-	ansible-playbook --extra-vars="@config.yml" -i inventory homelabos.yml
+deploy: logo
+	@ansible-playbook --extra-vars="@config.yml" -i inventory homelabos.yml
 
 logo:
-	cat homelaboslogo.txt
-
-get_roles:
-	ansible-galaxy --roles-path ./roles install toke.tor
+	@cat homelaboslogo.txt
 
 # Initial configuration
 config: logo
-	ansible-playbook --extra-vars="@config.yml" -i setup_inventory setup.yml
-	echo "========== Configuration completed! Now just run 'make' =========="
+# If config.yml does not exist, populate it with a 'blank'
+# yml file so the first attempt at parsing it succeeds
+	@[ -f config.yml ] || cp config.yml.blank config.yml
+	@ansible-playbook --extra-vars="@config.yml" -i setup_inventory setup.yml
+	@echo "========== Configuration completed! Now just run 'make' =========="
 
 # Reset all local settings
 config_reset: logo
-	cp config.yml.blank config.yml
+	@cp config.yml.blank config.yml
+	@echo "========== Configuration reset! Now just run 'make config' =========="
 
 # Update just HomelabOS Services (skipping slower initial setup steps)
 update: logo
-	ansible-playbook --extra-vars="@config.yml" -i inventory -t deploy homelabos.yml
+	@ansible-playbook --extra-vars="@config.yml" -i inventory -t deploy homelabos.yml
+	@echo "========== Update completed! =========="
 
 # Build the HomelabOs Documentation - Requires mkdocs with the Material Theme
 docs_build: logo
@@ -36,7 +38,7 @@ restore: logo
 	ansible-playbook -i inventory restore.yml
 
 # Spin up a development stack
-develop: logo get_roles
+develop: logo
 	vagrant plugin install vagrant-disksize
 	vagrant up
 	vagrant provision
